moved code out of main and into slack pkg
